requirements.txt
Package                Version
---------------------- ------------
asgiref                3.7.2
beautifulsoup4         4.12.2
bs4                    0.0.1
certifi                2023.7.22
cffi                   1.16.0
charset-normalizer     3.3.0
cryptography           41.0.4
Deprecated             1.2.14
Django                 4.2.6
django-oauth-toolkit   2.3.0
djangorestframework    3.14.0
idna                   3.4
jwcrypto               1.5.0
mysql-connector-python 8.1.0
oauthlib               3.2.2
pip                    23.3
protobuf               4.21.12
pycparser              2.21
pytz                   2023.3.post1
requests               2.31.0
setuptools             65.5.0
soupsieve              2.5
sqlparse               0.4.4
typing_extensions      4.8.0
urllib3                2.0.6
wrapt                  1.15.0
